refactor: move EIP-2537 and EIP-152 precompiles to shared hooks location
Move BLS12-381 (EIP-2537) and BLAKE2F (EIP-152) precompile
implementations from the Ethereum-specific hooks directory to a shared
bootloader-level hooks directory, so they can be used by both ZK and
Ethereum STFs. The Ethereum hooks module now re-exports from the shared
location behind feature gates, and the Ethereum post_init_op is updated
to use feature-gated initialization calls.
