+++ basic_bootloader/src/bootloader/hooks/eip_152/impls.rs
