+++ basic_bootloader/src/bootloader/hooks/eip_152/mixing_function.rs
