+++ basic_bootloader/src/bootloader/hooks/eip_2537/addresses.rs
